pyproject.toml
[project]
name = "nanosiglip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "aleks", email = "podidiving@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "safetensors>=0.7.0",
    "sentencepiece>=0.2.1",
    "torch>=2.10.0",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "protobuf>=7.34.0",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
    "transformers>=5.3.0",
]
